feat: Move config and print functions to separate files
Move the config file reading/creation and the print functions to
separate files to improve code structure.
